Pull out models into separate package
