+++ core/src/main/scala/boon/SuiteLike.scala
@@ -1,5 +1,7 @@
 package boon
 
+import boon.model._
+
 abstract class SuiteLike(val suiteName: String) {
 
   def tests: NonEmptySeq[DeferredTest]
