+++ core/src/main/scala/boon/model/Defer.scala
@@ -1,0 +1,10 @@
+package boon
+package model
+
+final case class Defer[A](value: () => A) {
+  def map[B](f: A => B): Defer[B] = Defer(() => f(value()))
+
+  def flatMap[B](f: A => Defer[B]): Defer[B] = Defer(() => f(value()).value())
+
+  def run(): A = value()
+}
