+++ core/src/main/scala/boon/model/Passable.scala
@@ -1,0 +1,12 @@
+package boon
+package model
+
+sealed trait Passable
+case object Passed extends Passable
+case object Failed extends Passable
+
+object Passable {
+  def hasPassed(passable: Passable): Boolean = passable == Passed
+
+  def hasFailed(passable: Passable): Boolean = !hasPassed(passable)
+}
