+++ core/src/main/scala/boon/model/Testable.scala
@@ -1,4 +1,5 @@
 package boon
+package model
 
 trait Testable {
   type Actual
