+++ core/src/main/scala/boon/syntax/StringRepSyntax.scala
@@ -1,6 +1,8 @@
 package boon
 package syntax
 
+import boon.model.StringRep
+
 final case class StringRepSyntax[T](value: T) {
   def strRep(implicit S: StringRep[T]): String = StringRep[T].strRep(value)
 }
